requirements.txt
deltalake
django
djangorestframework
streamlit
pandas
numpy
plotly